crawed each version's LOC

+++ PythonScripts/downloadPatches.py
@@ -5,6 +5,7 @@ from spider import getAllFilesUrls
 urls = getAllFilesUrls()
 print(urls)
 
+#download linux-kernel patch 2.6-4.16
 for url in urls:
     file_name = url.split('/')[-1]
     directory_name = '..\\kernel\\linux-' + '.'.join(file_name[6:].split('.')[:2]) + '\\'
@@ -13,9 +14,10 @@ for url in urls:
     if not os.path.exists(directory_name):
         os.makedirs(directory_name)
 
-    local_filename, headers = urllib.request.urlretrieve(url, local_file)
+    if not os.path.exists(local_file):
+        local_filename, headers = urllib.request.urlretrieve(url, local_file)
 
-    print(local_filename)
+    print(local_file)
 
 walk_dir_unzip('..\\kernel')
